feat(265): Custom AUR Helper
Instead of checking for installed AUR helpers liek yay and paru and then deciding what to do, this script initially asks the user what their preferred AUR helper is, it then checks if the chosen helper is installed. If not, the script then installs the chosen AUR helper and cleans up after it self at then end.