dotfiles: Fix auto-complete on ls, ll and lt by passing --icons=always argument
The --icons argument in eza command expects either "always, auto,
automatic or never" to be passed, hence causing the auto complete on
hitting tab key to never list the directories after ls, ll and lt
command. Now pass "always" to it so that the aesthetics are maintained
and get auto-complete to list the directories and files as intended.

Change-Id: Ic4775d0daa59de21b06e01dc9d77e4e7f72c3907